requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
falcon==2.0.0
hug==2.6.1
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
requests==2.28.1
urllib3==1.26.13

DB_HOST = k3database.cjbgtbz37j5g.ap-southeast-3.rds.amazonaws.com
DB_USER = admin
DB_PASSWORD = admin123
DB_NAME = k3database
DB_PORT = 3306